backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
faster-whisper
anyio
numpy
soundfile
ffmpeg-python
google-generativeai
python-dotenv
